// backend/auto-operation/proto/state/v1/block_pb.ts
export enum BlockStateEnum {
  /**
   * @generated from enum value: BLOCK_STATE_UNKNOWN = 0;
   */
  BLOCK_STATE_UNKNOWN = 0,

  /**
   * 閉塞が開の状態(列車がいない)
   *
   * @generated from enum value: BLOCK_STATE_OPEN = 1;
   */
  BLOCK_STATE_OPEN = 1,

  /**
   * 閉塞が閉の状態(列車がいない)
   *
   * @generated from enum value: BLOCK_STATE_CLOSE = 2;
   */
  BLOCK_STATE_CLOSE = 2,
}